# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Zydesk API",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/api"
    }
  ],
  "paths": {
...
    "/auth/ingresar": {
      "post": {
        "summary": "Ingresar con correo y contraseña",
        "description": "Permiso: publico",
        "tags": [
          "Autenticación"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "correo": {
                    "type": "string",
                    "maxLength": 200,
                    "format": "email",
                    "pattern": "^(?:[A-Za-z0-9_'+\\-]+\\.)*[A-Za-z0-9_'+\\-]*[A-Za-z0-9_+-]@(?:[A-Za-z0-9][A-Za-z0-9\\-]*\\.)+[A-Za-z]{2,}$"
                  },
                  "contrasena": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "mantener": {
                    "default": false,
                    "type": "boolean"
                  }
                },
                "required": [
                  "correo",
                  "contrasena"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Respuesta correcta",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "exclusiveMinimum": 0,
                      "maximum": 9007199254740991
                    },
                    "nombre": {
                      "type": "string"
                    },
                    "correo": {
                      "type": "string"
                    },
                    "rol": {
                      "type": "string",
                      "enum": [
                        "admin",
                        "coordinacion",
                        "tecnico",
                        "lectura"
                      ]
                    },
                    "departamento": {
                      "anyOf": [
                        {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "integer",
                              "exclusiveMinimum": 0,
                              "maximum": 9007199254740991
                            },
                            "nombre": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "id",
                            "nombre"
                          ],
                          "additionalProperties": false
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "color_avatar": {
                      "type": "string"
                    },
                    "iniciales": {
                      "type": "string"
                    },
                    "permisos": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "enum": [
                          "tickets.editar",
                          "ots.aprobar",
                          "ots.cerrar",
                          "ots.facturar",
                          "reportes.ver",
                          "config.editar"
                        ]
                      }
                    },
                    "debe_cambiar_contrasena": {
                      "type": "boolean"
                    },
                    "debe_aceptar_terminos": {
                      "type": "boolean"
                    },
                    "terminos_version_vigente": {
                      "type": "string"
                    },
                    "nombre_app": {
                      "type": "string"
                    },
                    "logo_url": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  },
                  "required": [
                    "id",
                    "nombre",
                    "correo",
                    "rol",
                    "departamento",
                    "color_avatar",
                    "iniciales",
                    "permisos",
                    "debe_cambiar_contrasena",
                    "debe_aceptar_terminos",
                    "terminos_version_vigente",
                    "nombre_app",
                    "logo_url"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "codigo": {
                          "type": "string"
                        },
                        "mensaje": {
                          "type": "string"
                        },
                        "detalles": {}
                      },
                      "required": [
                        "codigo",
                        "mensaje"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "Conflicto",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "codigo": {
                          "type": "string"
                        },
                        "mensaje": {
                          "type": "string"
                        },
                        "detalles": {}
                      },
                      "required": [
                        "codigo",
                        "mensaje"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/auth/salir": {
      "post": {
        "summary": "Cerrar la sesión actual",
        "description": "Permiso: sesion",
        "tags": [
          "Autenticación"
        ],
        "responses": {
          "204": {
            "description": "Sin contenido"
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "codigo": {
                          "type": "string"
                        },
                        "mensaje": {
                          "type": "string"
                        },
                        "detalles": {}
                      },
                      "required": [
                        "codigo",
                        "mensaje"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "codigo": {
                          "type": "string"
                        },
                        "mensaje": {
                          "type": "string"
                        },
                        "detalles": {}
                      },
                      "required": [
                        "codigo",
                        "mensaje"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "Conflicto",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "codigo": {
                          "type": "string"
                        },
                        "mensaje": {
                          "type": "string"
                        },
                        "detalles": {}
                      },
                      "required": [
                        "codigo",
                        "mensaje"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/yo": {
      "get": {
        "summary": "Datos de la persona autenticada",
        "description": "Permiso: sesion",
        "tags": [
          "Yo"
        ],
        "responses": {
          "200": {
            "description": "Respuesta correcta",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "exclusiveMinimum": 0,
                      "maximum": 9007199254740991
                    },
                    "nombre": {
                      "type": "string"
                    },
                    "correo": {
                      "type": "string"
                    },
                    "rol": {
                      "type": "string",
                      "enum": [
                        "admin",
                        "coordinacion",
                        "tecnico",
                        "lectura"
                      ]
                    },
                    "departamento": {
                      "anyOf": [
                        {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "integer",
                              "exclusiveMinimum": 0,
                              "maximum": 9007199254740991
                            },
                            "nombre": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "id",
                            "nombre"
                          ],
                          "additionalProperties": false
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "color_avatar": {
                      "type": "string"
                    },
                    "iniciales": {
                      "type": "string"
                    },
                    "permisos": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "enum": [
                          "tickets.editar",
                          "ots.aprobar",
                          "ots.cerrar",
                          "ots.facturar",
                          "reportes.ver",
                          "config.editar"
                        ]
                      }
                    },
                    "debe_cambiar_contrasena": {
                      "type": "boolean"
                    },
                    "debe_aceptar_terminos": {
                      "type": "boolean"
                    },
                    "terminos_version_vigente": {
                      "type": "string"
                    },
                    "nombre_app": {
                      "type": "string"
                    },
                    "logo_url": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  },
                  "required": [
                    "id",
                    "nombre",
                    "correo",
                    "rol",
                    "departamento",
                    "color_avatar",
                    "iniciales",
                    "permisos",
                    "debe_cambiar_contrasena",
                    "debe_aceptar_terminos",
                    "terminos_version_vigente",
                    "nombre_app",
                    "logo_url"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "codigo": {
                          "type": "string"
                        },
                        "mensaje": {
                          "type": "string"
                        },
                        "detalles": {}
                      },
                      "required": [
                        "codigo",
                        "mensaje"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "codigo": {
                          "type": "string"
                        },
                        "mensaje": {
                          "type": "string"
                        },
                        "detalles": {}
                      },
                      "required": [
                        "codigo",
                        "mensaje"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/yo/aceptar-terminos": {
      "post": {
        "summary": "Aceptar la versión vigente de los términos",
        "description": "Permiso: sesion",
        "tags": [
          "Yo"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "version": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "version"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Respuesta correcta",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "exclusiveMinimum": 0,
                      "maximum": 9007199254740991
                    },
                    "nombre": {
                      "type": "string"
                    },
                    "correo": {
                      "type": "string"
                    },
                    "rol": {
                      "type": "string",
                      "enum": [
                        "admin",
                        "coordinacion",
                        "tecnico",
                        "lectura"
                      ]
                    },
                    "departamento": {
                      "anyOf": [
                        {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "integer",
                              "exclusiveMinimum": 0,
                              "maximum": 9007199254740991
                            },
                            "nombre": {
                              "type": "string"
                            }
                          },
                          "required": [
                            "id",
                            "nombre"
                          ],
                          "additionalProperties": false
                        },
                        {
                          "type": "null"
                        }
                      ]
                    },
                    "color_avatar": {
                      "type": "string"
                    },
                    "iniciales": {
                      "type": "string"
                    },
                    "permisos": {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "enum": [
                          "tickets.editar",
                          "ots.aprobar",
                          "ots.cerrar",
                          "ots.facturar",
                          "reportes.ver",
                          "config.editar"
                        ]
                      }
                    },
                    "debe_cambiar_contrasena": {
                      "type": "boolean"
                    },
                    "debe_aceptar_terminos": {
                      "type": "boolean"
                    },
                    "terminos_version_vigente": {
                      "type": "string"
                    },
                    "nombre_app": {
                      "type": "string"
                    },
                    "logo_url": {
                      "type": [
                        "string",
                        "null"
                      ]
                    }
                  },
                  "required": [
                    "id",
                    "nombre",
                    "correo",
                    "rol",
                    "departamento",
                    "color_avatar",
                    "iniciales",
                    "permisos",
                    "debe_cambiar_contrasena",
                    "debe_aceptar_terminos",
                    "terminos_version_vigente",
                    "nombre_app",
                    "logo_url"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Datos inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "codigo": {
                          "type": "string"
                        },
                        "mensaje": {
                          "type": "string"
                        },
                        "detalles": {}
                      },
                      "required": [
                        "codigo",
                        "mensaje"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "codigo": {
                          "type": "string"
                        },
                        "mensaje": {
                          "type": "string"
                        },
                        "detalles": {}
                      },
                      "required": [
                        "codigo",
                        "mensaje"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "Sin permiso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "codigo": {
                          "type": "string"
                        },
                        "mensaje": {
                          "type": "string"
                        },
                        "detalles": {}
                      },
                      "required": [
                        "codigo",
                        "mensaje"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "409": {
            "description": "Conflicto",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "codigo": {
                          "type": "string"
                        },
                        "mensaje": {
                          "type": "string"
                        },
                        "detalles": {}
                      },
                      "required": [
                        "codigo",
                        "mensaje"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sesion"
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}